Clean up HTML code and complete implementation for limit
Cleaned up the HTML code and implemented a text box where the user can limit the amount of todos that appear. The "limit" filter is now done on the client side to prevent bugs with limiting the wrong way (it would limit before adding other filters and therefore wouldn't give all available options). Modified testing in todos.service to accommodate for this.